Moved Message and Emote tsx to Generics and Deleted Message.scss and Emote.scss
