+++ src/javascript/components/Chat/ChatEventsPopup.tsx
@@ -3,7 +3,6 @@ import { useContext, Component, useState, useEffect } from 'react';
 import { theme, ThemeContext } from '../../helpers';
 import { MdSend, MdPerson, MdMood, MdFace, MdPlayArrow } from 'react-icons/md';
 
-import { Message } from './Message';
 import { firebaseConfig$, setRxConfig } from '../../helpers/rxConfig';
 import { Action } from 'rxjs/internal/scheduler/Action';
 
@@ -27,11 +26,7 @@ import { first } from 'rxjs/operators';
 const Window: any = window;
 const { ipcRenderer, shell } = Window.require('electron');
 
-const styles: any = require('./Chat.scss');
-const segStyles: any = require('../SegmentControl/SegmentControl.scss');
-
 interface popup {
-  styles: any;
   stateTheme: any;
   text?: string | Function | Element | any;
   Config?: any;
@@ -48,7 +43,6 @@ let ninjaTimeout;
 let ninjetTimeout;
 
 const ChatEventsPopup = ({
-  styles,
   stateTheme,
   text = '',
   Config = {},
