--- src/javascript/components/Chat/Message.scss
@@ -1,3 +1,0 @@
-.highlighted {
-  border: 2px solid rgba(255, 211, 0, 0.6);
-}
